Adds Data class, as well as Touchable class.  Data holds data points.  Touchable creates new motion event

+++ MyApplication/app/src/main/java/ubhacking/thumbpedometer/Data.java
@@ -4,7 +4,7 @@ package ubhacking.thumbpedometer;
  * Created by John on 11/7/14.
  */
 public class Data {
-    private int _xDist, _yDist, _totalDist;
+    private float _xDist, _yDist, _totalDist;
 
     public Data(){
